# swagger.yaml
openapi: 3.0.3
info:
    title: GitHub Release Notification API
    version: 1.0.0
    description: API for subscribing to email notifications about new GitHub repository releases.
servers:
    - url: http://localhost:3000
...
paths:
...
    /api/subscriptions:
        get:
            tags:
                - Subscriptions
            summary: Get all active subscriptions for an email
            description: Returns all active repository subscriptions associated with the given email.
            operationId: getSubscriptionsByEmail
            parameters:
                - name: email
                  in: query
                  required: true
                  description: Subscriber email
                  schema:
                      type: string
                      format: email
            responses:
                "200":
                    description: Active subscriptions retrieved successfully.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SubscriptionsResponse"
                "400":
                    description: Invalid email query parameter.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: No subscriptions found for this email.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
components:
    schemas:
...
        SubscriptionItem:
            type: object
            additionalProperties: false
            required:
                - repository
                - status
                - createdAt
            properties:
                repository:
                    type: string
                    example: golang/go
                status:
                    type: string
                    enum:
                        - active
                createdAt:
                    type: string
                    format: date-time
                    example: 2026-04-08T10:30:00Z
        SubscriptionsResponse:
            type: object
            additionalProperties: false
            required:
                - email
                - subscriptions
            properties:
                email:
                    type: string
                    format: email
                    example: user@example.com
                subscriptions:
                    type: array
                    items:
                        $ref: "#/components/schemas/SubscriptionItem"
        ErrorResponse:
            type: object
            additionalProperties: false
            required:
                - error
                - message
            properties:
                error:
                    type: string
                    example: Bad Request
                message:
                    type: string
                    example: Repository format must be owner/repo.